ADDON CHANGES
This addon no longer has the option to create tags since this wasn't functioning as intended. Unless someone could help with finding the proper settings like rotations and whatever the tags need, this function will stay away.
Only option you have is parent all objects and tags to eachother in 1 button (all lods at once)
And it has a button to clean the hierarchy, this will be explained in my jkhub tutorial.