Update __init__.py
Changed the way of searching for where to parent to. It now has some failsafes if multiple objects like head_eyes and torso_shoulder etc are in the mix. These will be parented to its respective part. stupidtriangle and stupidtriangle_off are now supported aswell.

The code looks a little messy but as soon as I learn better techniques, that will be fixed.