3.1.3. Piece selector and rework
In this version, I have added a function so you can select Meshes, Tags and Caps separately, which helps on the ports.
I have redone the parenting function to admit parenting even if convention is not followed, which will be assigned by default to "hips_LOD".
Also introduced a new change into the model_default.skin creator, so that if you have selected a special termination for your texture, it will assign this one (has to be first material slot) otherwise it will assign .tga