Utility update
Added functionality for tags to be created for each model_root there currently is in the scene.

E.g.: model_root_0 and model_root_1 is currently in the scene, then it will create tags for 0 and 1. These must be parented with any of the buttons created for that.

Also, renaming Blender Import Export Plugin from MrWonko no longer has to be renamed. So that issue is resolved aswell.