Functionality added
Added the option to decide what to do with the replaced object.

DELETE: deletes object 1, renames object 2 and reparents children of object 1 to object 2.
UNPARENT: unparents object 1 and renames it to replaced_name it had before. replace children's parenting from object 1 to object 2.
KEEP: unparents the children of object 1 and put them on object 2. Rename object 2 to what object 1 was while object 1 gets the name replaced_name it was before